refactor project structure: separate Navbar into its own component and file
