+++ src/pages/Contact-Us/contact.us.css
@@ -1,0 +1,7 @@
+.contact-us-page{
+    min-height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background: linear-gradient(180deg, #030318 0%, #0a0f3d 45%, #121b6e 100%);
+}
