+++ src/pages/navbar/navbar.css
@@ -1,0 +1,374 @@
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
+
+body {
+    font-family: Arial, sans-serif;
+    background: linear-gradient(180deg, #030318 0%, #0a0f3d 40%, #332fd0 100%);
+    color: white;
+}
+
+.app {
+    width: 100%;
+    min-height: 100vh;
+}
+
+#network-canvas {
+    position: fixed;
+    top: 0;
+    left: 0;
+    z-index: -1;
+}
+
+.navbar {
+    position: relative;
+    z-index: 10;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 30px 60px;
+}
+
+.logo {
+    font-size: 22px;
+    font-weight: bold;
+    cursor: pointer;
+    color: white;
+    
+}
+
+.navbar-buttons{
+    text-decoration: none;
+}
+
+.nav-links {
+    display: flex;
+    gap: 25px;
+    align-items: center;
+}
+
+.nav-link {
+    color: white;
+    text-decoration: none;
+    font-size: 14px;
+}
+
+.nav-btn {
+    background: none;
+    border: none;
+    color: white;
+    font-size: 14px;
+    cursor: pointer;
+}
+
+.nav-btn,
+.nav-link {
+    position: relative;
+}
+
+.nav-btn::after,
+.nav-link::after {
+    content: "";
+    position: absolute;
+    left: 0;
+    bottom: -6px;
+    width: 0%;
+    height: 2px;
+    background: white;
+    transition: width 0.3s ease;
+}
+
+.nav-btn:hover::after,
+.nav-link:hover::after {
+    width: 100%;
+}
+
+.dropdown {
+    position: relative;
+}
+
+.dropdown-content {
+    position: absolute;
+    top: 28px;
+    left: 0;
+    background: #141453;
+    min-width: 180px;
+    border-radius: 8px;
+    padding: 8px 0;
+}
+
+.dropdown-content a {
+    display: block;
+    padding: 10px 15px;
+    color: white;
+    text-decoration: none;
+}
+
+.dropdown-content a:hover {
+    background: #3a32c9;
+}
+
+.contact-btn {
+    position: relative;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 190px;
+    height: 44px;
+    padding: 0 20px;
+    border-radius: 40px;
+    border: none;
+    background: white;
+    cursor: pointer;
+    overflow: hidden;
+    font-weight: 600;
+}
+
+
+/* TEXT (STATIC – NO ANIMATION) */
+
+.contact-btn .text {
+    transition: opacity 0.2s ease;
+    color: #000;
+}
+
+
+/* ICON */
+
+.contact-btn .icon {
+    position: absolute;
+    right: 20px;
+    font-size: 18px;
+    transition: transform 0.45s ease;
+}
+
+
+/* ON HOVER */
+
+.contact-btn:hover .text {
+    opacity: 0;
+    /* text disappears */
+}
+
+.contact-btn:hover .icon {
+    transform: translateX(-130px);
+    /* emoji slides right → left */
+}
+
+
+/* Mobile tap feedback */
+
+.contact-btn:active {
+    transform: scale(0.96);
+
+    display: flex;
+    align-items: center;
+    background: white;
+    border-radius: 50px;
+    padding: 6px 8px 6px 20px;
+    color: black;
+    font-size: 14px;
+    font-weight: 600;
+    cursor: pointer;
+    gap: 12px;
+
+}
+
+.btn-text {
+    letter-spacing: 1px;
+}
+
+.contact-icon {
+    width: 34px;
+    height: 34px;
+    background: #2563eb;
+    border-radius: 50%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.hero-content {
+    position: relative;
+    z-index: 2;
+    height: 100vh;
+    padding-left: 80px;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+}
+
+.hero-content h1 {
+    font-size: 48px;
+}
+
+.hero-content h2 {
+    font-size: 52px;
+    margin: 10px 0;
+}
+
+.hero-content p {
+    margin-top: 10px;
+    opacity: 0.8;
+}
+
+.link-btn {
+    background: transparent;
+    border: none;
+    color: white;
+    width: 100%;
+    text-align: left;
+    padding: 10px 16px;
+    font-size: 14px;
+    cursor: pointer;
+    position: relative;
+}
+
+.link-btn:hover {
+    background: rgba(255, 255, 255, 0.08);
+}
+
+.link-btn::after {
+    content: "";
+    position: absolute;
+    left: 16px;
+    bottom: 6px;
+    width: 0%;
+    height: 1px;
+    background: white;
+    transition: width 0.25s ease;
+}
+
+.link-btn:hover::after {
+    width: calc(100% - 32px);
+}
+
+.nav-link-btn {
+    background: none;
+    border: none;
+    color: white;
+    font-size: 14px;
+    font-weight: 500;
+    cursor: pointer;
+    padding: 4px 0;
+    position: relative;
+}
+
+.nav-link-btn::after {
+    content: "";
+    position: absolute;
+    left: 0;
+    bottom: -6px;
+    width: 0%;
+    height: 2px;
+    background: white;
+    transition: width 0.3s ease;
+}
+
+.nav-link-btn:hover::after {
+    width: 100%;
+}
+
+.contact-wrapper {
+    background: #f4f6f8;
+    color: black;
+    padding: 80px 20px;
+}
+
+.hamburger {
+    width: 32px;
+    height: 24px;
+    display: none;
+    flex-direction: column;
+    justify-content: space-between;
+    cursor: pointer;
+}
+
+.hamburger span {
+    height: 3px;
+    width: 100%;
+    background: white;
+    border-radius: 3px;
+    transition: all 0.35s ease;
+}
+
+.hamburger.open span:nth-child(1) {
+    transform: rotate(45deg) translate(6px, 6px);
+}
+
+.hamburger.open span:nth-child(2) {
+    opacity: 0;
+}
+
+.hamburger.open span:nth-child(3) {
+    transform: rotate(-45deg) translate(6px, -6px);
+}
+
+@media (max-width: 768px) {
+    .hamburger {
+        display: flex;
+    }
+    .navbar {
+        padding: 20px;
+    }
+    .nav-links {
+        position: absolute;
+        top: 70px;
+        left: 0;
+        width: 100%;
+        background: rgba(11, 11, 42, 0.95);
+        flex-direction: column;
+        align-items: flex-start;
+        padding: 20px;
+        display: none;
+    }
+    .nav-links.open {
+        display: flex;
+    }
+    .dropdown-content {
+        position: static;
+        width: 100%;
+        margin-top: 10px;
+    }
+    .hero-content {
+        padding: 0 20px;
+        text-align: center;
+    }
+    .hero-content h1 {
+        font-size: 28px;
+    }
+    .hero-content h2 {
+        font-size: 32px;
+    }
+    .hero-content p {
+        font-size: 14px;
+    }
+    .contact-btn {
+        position: fixed;
+        bottom: 16px;
+        left: 50%;
+        transform: translateX(-50%);
+        width: calc(100% - 32px);
+        max-width: 420px;
+        padding: 12px 18px;
+        border-radius: 999px;
+        justify-content: center;
+        gap: 12px;
+        z-index: 2000;
+        box-shadow: 0 10px 30px rgba(0, 0, 0, 0.35);
+    }
+    .contact-btn .btn-text {
+        font-size: 15px;
+        letter-spacing: 1px;
+    }
+    .contact-icon {
+        width: 40px;
+        height: 40px;
+    }
+    .contact-icon span,
+    .contact-icon svg {
+        font-size: 18px;
+    }
+}
